# doc/api.yaml
openapi: 3.0.3
info:
  title: UniDrive - OpenAPI 3.0
  description: |-
    Share part of the whole journey to the University with other students thanks to UniDrive!
    
  contact:
    email: vecchi.1938069@studenti.uniroma1.it
  license:
    name: Sapienza University
    url: https://www.uniroma1.it/it/
  version: 1.0.0
...
servers:
  - url: '{protocol}://UniDrive.com'
    variables:
      protocol:
        enum:
          - http
          - https
        default: https
...
paths:
...
  /profile/{user_id}:
    parameters:
      - name: "user_id"
        in: path
        required: true
        schema: 
          $ref: '#/components/schemas/user_id'
    get:
      tags:
          - Profile
      summary: visualize an user profile
      operationId: getProfile
      description: View the profile of another user
      responses:
        '200':
          description: The user's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '404': { description: User not found }
        
        '500': {$ref: "#/components/responses/InternalServerError" }
    put:
      tags:
          - Profile
      summary: update an user profile
      operationId: updateProfile
      description:  Update the user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
      responses:
        '200':
          description: Confirmation of the update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'

        '500': {$ref: "#/components/responses/InternalServerError" }
...
  /profile/{user_id}/review:
    parameters:
    - name: user_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/user_id'
    
    post:
      tags:
        - Review
      summary: create a review
      operationId: createReview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'

      responses:
        '201':
          description: create a review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
    
    get: 
      tags:
        - Review
      summary: get reviews
      operationId: getReviews
      responses:
        '200':
          description: The review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
...
components:
...
  schemas:
    depart_hour:
      description: When we want to book the ride
      type: string
      format: date-time
      example: "2022-11-04 19:35"
      pattern: '^20[2-9][0-9]-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01]) (([0-1][0-9]|2[0-3]):[0-5][0-9])$'
    Ride_id:
      description: Unique identifier representing a specific ride
      type: string
      pattern: '^[a-zA-Z0-9-_]*$'
      minLength: 1
      maxLength: 20
      example: "U3dhZ2dlciByb2Nrcw"
...
    user_id:
      description: Unique identifier representing a specific user
      type: string
      pattern: '^[a-zA-Z0-9-_]*$'
      minLength: 1
      maxLength: 20
      example: "U3dhZ2dlciByb2Nrcw"
    username:
      type: string
      example: Mario Rossi
      pattern: '^[a-zA-Z0-9-_]*$'
      minLength: 3
      maxLength: 16
    name:
      type: string
      example: Mario 
      pattern: '^[a-zA-Z-_]*$'
      minLength: 3
      maxLength: 16
    surname:
      type: string
      example: Rossi
      pattern: '^[a-zA-Z-_]*$'
      minLength: 3
      maxLength: 16
    Ride:
      type: object
      properties:
        id: 
          $ref: '#/components/schemas/Ride_id'
        origin:
          description: The starting point of the ride
          type: string
          example: Via di Valle Giulia, 9, 00197 Roma RM
        destination:
          description: The end point of the ride
          type: string
          example: Piazzale Aldo Moro, 5, 00185 Roma RM
        driver_name:
          description: The name of the user who is offering the ride
          $ref: '#/components/schemas/username'
        driver_id:
          description: the id of the driver
          $ref: '#/components/schemas/user_id'
        available_seats:
          description: The number of seats available in the car
          type: integer
          example: 3
        depart-hour:
          $ref: '#/components/schemas/depart_hour'
...
    Profile:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/user_id'
        name:
          $ref: '#/components/schemas/name'
        surname:
          $ref: '#/components/schemas/surname'
        age:
          description: Age of the user
          type: integer
          example: 21
          minimum: 18
        faculty:
          description: Faculty of the university that the user belongs to
          type: string
          example: Faculty of Economics
        profile_picture_url:
          description: URL of the profile picture. Accepting only ID attached with.png/.jpg/.jpeg extensions.
          type: string
          pattern: '^[a-zA-Z0-9-_]*\.(png|jpg|jpeg)$'
          example: "Cksvy-ZXtXRE.jpg"
          minLength: 0
          maxLength: 150
        bio:
          description: Biography of the profile. |
            Keeping the pattern simple. Not doing crazy things. Just allowing alphanumeric characters and basic punctuation.
          type: string
          pattern: '^[a-zA-Z0-9,._:;?!\x27\- ]*$'
          example: "Mario, business student at Sapienza"
          minLength: 0
          maxLength: 150
        reviews:
          description: List of reviews received by the user
          type: array
          items:
            $ref: '#/components/schemas/Review'
        #rides_booked:
          #description: List of rides booked by the user
         # type: array
         # items:
         #   $ref: '#/components/schemas/Ride'
        total_rides_given:
          description: Total number of rides given by the user
          type: integer
          example: 50
          minimum: 0
        joined_in:
          description: When the user joined the platform
          type: string
          pattern: '^(?i)(january|february|march|april|may|june|july|august|september|october|november|december)\s\d{4}$'
          example: September 2022
          minLength: 8
          maxLength: 14
        car_details:
          $ref: '#/components/schemas/Car_details'
        city:
          description: The place where the user lives
          type: string
          example: Rome
          pattern: '^[a-zA-Z-_]*$'
          minLength: 0
          maxLength: 20
        district:
          description: The district where the user lives
          type: string
          example: Parioli
          pattern: '^[a-zA-Z-_]*$'
          minLength: 0
          maxLength: 30
        #social_links:
        #  description: List of social links of the user
        #  type: array
        #  items:
        #    $ref: '#/components/schemas/SocialLink'
        instagram_url:
          description: The url of the social link
          type: string
          example: https://www.instagram.com/mario_rossi/
          pattern: '^(http|https):\/\/(www\.)?([a-zA-Z0-9-_]+)\.([a-zA-Z0-9-_]+)\/([a-zA-Z0-9-_]+)\/?$'
          minLength: 0
          maxLength: 150
    Car_details:
      type: object
      properties:
        car_model:
          description: The model of the car
          type: string
          example: Fiat 500
          pattern: '^[a-zA-Z0-9-_]*$'
          minLength: 2
          maxLength: 20
        car_color:
          description: The color of the car
          type: string
          example: Red
          pattern: '^[a-zA-Z-_]*$'
          minLength: 2
          maxLength: 20
        car_plate:
          description: The plate of the car
          type: string
          example: "AB123CD"
          pattern: '^[A-Z]{2}[0-9]{3}[A-Z]{2}$'
          minLength: 0
          maxLength: 7
...
    SocialLink:
      type: object
      properties:
        name:
          description: The name of the social link
          type: string
          example: Instagram
          pattern: '^[a-zA-Z0-9-_]*$'
          minLength: 3
          maxLength: 16
        url:
          description: The url of the social link
          type: string
          example: https://www.instagram.com/mario_rossi/
          pattern: '^(http|https):\/\/(www\.)?([a-zA-Z0-9-_]+)\.([a-zA-Z0-9-_]+)\/([a-zA-Z0-9-_]+)\/?$'
          minLength: 0
          maxLength: 150
    Review:
      type: object
      properties:
        user_id:
          $ref: '#/components/schemas/user_id'
        rating:
          description: The rating of the user
          type: integer
          example: 4
          minimum: 0
          maximum: 5
        review:
          description: The review of the user
          type: string
          example: "Mario is a great driver"
          pattern: '^[a-zA-Z0-9,._:;?!\x27\- ]*$'
          minLength: 0
          maxLength: 300
        timestamp:
          description: When the review was written
          type: string
          format: date-time
          example: 2022-11-04T19:35:22-04:00
          minLength: 20
          maxLength: 25
        author_name:
          description: The name of the user who wrote the review
          type: string
          example: John Doe
          pattern: '^[a-zA-Z0-9-_]*$'
          minLength: 3
          maxLength: 16
...
  responses:
...
    InternalServerError: 
      description: |-
        The server encountered an internal error.
        Further info in server logs.
...